incremental mapper impl
